+++ CHANGELOG.rst
@@ -9,6 +9,15 @@ v0.1.0
 
 New features
 ~~~~~~~~~~~~
+* Added ``-V``, ``--version`` arguments to print program's version information.
+* Added ``-dumpversion`` argument to print the version of the program and don't
+  do anything else.
+
+v0.1.0
+------
+
+New features
+~~~~~~~~~~~~
 
 * The GitHub username is no longer required upon obtaining organizations list.
 * Provided ability to pass authentication token for github.com API requests via
@@ -18,7 +27,7 @@ New features
 Packaging changes
 ~~~~~~~~~~~~~~~~~
 
-* Change the way to communicate with GitHub API. ```requests`` library no
+* Changed the way to communicate with GitHub API. ``requests`` library no
   longer used thanks to ``PyGithub``.
 
 v0.0.5